vrrp: extract GARP/gateway-probe to instance_garp.go
Pure code-motion split of pkg/vrrp/instance.go (#5661). Move the
gratuitous-ARP / unsolicited-NA burst path, its send-dampener gates,
and the gateway-probe target derivation into a cohesive sibling file,
package vrrp:

  minGARPInterval, garpDampened, garpSendAllowed, arpProbeFn,
  garpBurstFn/naBurstFn, GatewayProbeTarget, sendGARP.

Declarations moved verbatim — no rename of exported symbols (the
exported GatewayProbeTarget is unchanged and still reused by
pkg/daemon), no logic/timing/const change, and no reorder of any
goroutine, defer, channel, or mutex operation. The two suppression
gates (epoch dedup + 500ms time dampener) and their force-bypass
semantics are byte-identical. Imports pruned by goimports; gofmt
clean. Builds green.
